Moved all embed code to seperate folder
