+++ lib/flutter_quill.dart
@@ -1,5 +1,7 @@
 library flutter_quill;
 
+export 'src/embeds/default_embed_builder.dart';
+export 'src/embeds/widgets/image.dart';
 export 'src/models/documents/attribute.dart';
 export 'src/models/documents/document.dart';
 export 'src/models/documents/nodes/embeddable.dart';
@@ -14,7 +16,6 @@ export 'src/utils/embeds.dart';
 export 'src/widgets/controller.dart';
 export 'src/widgets/default_styles.dart';
 export 'src/widgets/editor.dart';
-export 'src/widgets/embeds/image.dart';
 export 'src/widgets/link.dart' show LinkActionPickerDelegate, LinkMenuAction;
 export 'src/widgets/style_widgets/style_widgets.dart';
 export 'src/widgets/toolbar.dart';
