+++ lib/src/embeds/default_embed_builder.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -6,18 +8,34 @@ import 'package:gallery_saver/gallery_saver.dart';
 import 'package:math_keyboard/math_keyboard.dart';
 import 'package:tuple/tuple.dart';
 
-import '../../models/documents/attribute.dart';
-import '../../models/documents/nodes/embeddable.dart';
-import '../../models/documents/nodes/leaf.dart' as leaf;
-import '../../translations/toolbar.i18n.dart';
-import '../../utils/embeds.dart';
-import '../../utils/platform.dart';
-import '../../utils/string.dart';
-import '../controller.dart';
-import 'image.dart';
-import 'image_resizer.dart';
-import 'video_app.dart';
-import 'youtube_video_app.dart';
+import '../models/documents/attribute.dart';
+import '../models/documents/nodes/embeddable.dart';
+import '../models/documents/nodes/leaf.dart' as leaf;
+import '../translations/toolbar.i18n.dart';
+import '../utils/embeds.dart';
+import '../utils/platform.dart';
+import '../utils/string.dart';
+import '../widgets/controller.dart';
+import 'toolbar/image_video_utils.dart';
+import 'widgets/image.dart';
+import 'widgets/image_resizer.dart';
+import 'widgets/video_app.dart';
+import 'widgets/youtube_video_app.dart';
+
+export 'toolbar/image_button.dart';
+export 'toolbar/image_video_utils.dart';
+export 'toolbar/video_button.dart';
+
+typedef OnImagePickCallback = Future<String?> Function(File file);
+typedef OnVideoPickCallback = Future<String?> Function(File file);
+typedef FilePickImpl = Future<String?> Function(BuildContext context);
+typedef WebImagePickImpl = Future<String?> Function(
+    OnImagePickCallback onImagePickCallback);
+typedef WebVideoPickImpl = Future<String?> Function(
+    OnVideoPickCallback onImagePickCallback);
+typedef MediaPickSettingSelector = Future<MediaPickSetting?> Function(
+    BuildContext context);
+
 
 Widget defaultEmbedBuilder(
   BuildContext context,
