+++ lib/src/embeds/toolbar/video_button.dart
@@ -4,18 +4,20 @@ import 'package:image_picker/image_picker.dart';
 import '../../models/documents/nodes/embeddable.dart';
 import '../../models/themes/quill_dialog_theme.dart';
 import '../../models/themes/quill_icon_theme.dart';
-import '../controller.dart';
-import '../toolbar.dart';
+import '../../widgets/controller.dart';
+import '../../widgets/toolbar.dart';
+import '../default_embed_builder.dart';
+import 'image_video_utils.dart';
 
-class ImageButton extends StatelessWidget {
-  const ImageButton({
+class VideoButton extends StatelessWidget {
+  const VideoButton({
     required this.icon,
     required this.controller,
     this.iconSize = kDefaultIconSize,
-    this.onImagePickCallback,
+    this.onVideoPickCallback,
     this.fillColor,
     this.filePickImpl,
-    this.webImagePickImpl,
+    this.webVideoPickImpl,
     this.mediaPickSettingSelector,
     this.iconTheme,
     this.dialogTheme,
@@ -29,9 +31,9 @@ class ImageButton extends StatelessWidget {
 
   final QuillController controller;
 
-  final OnImagePickCallback? onImagePickCallback;
+  final OnVideoPickCallback? onVideoPickCallback;
 
-  final WebImagePickImpl? webImagePickImpl;
+  final WebVideoPickImpl? webVideoPickImpl;
 
   final FilePickImpl? filePickImpl;
 
@@ -61,13 +63,13 @@ class ImageButton extends StatelessWidget {
   }
 
   Future<void> _onPressedHandler(BuildContext context) async {
-    if (onImagePickCallback != null) {
+    if (onVideoPickCallback != null) {
       final selector =
           mediaPickSettingSelector ?? ImageVideoUtils.selectMediaPickSetting;
       final source = await selector(context);
       if (source != null) {
         if (source == MediaPickSetting.Gallery) {
-          _pickImage(context);
+          _pickVideo(context);
         } else {
           _typeLink(context);
         }
@@ -77,13 +79,13 @@ class ImageButton extends StatelessWidget {
     }
   }
 
-  void _pickImage(BuildContext context) => ImageVideoUtils.handleImageButtonTap(
+  void _pickVideo(BuildContext context) => ImageVideoUtils.handleVideoButtonTap(
         context,
         controller,
         ImageSource.gallery,
-        onImagePickCallback!,
+        onVideoPickCallback!,
         filePickImpl: filePickImpl,
-        webImagePickImpl: webImagePickImpl,
+        webVideoPickImpl: webVideoPickImpl,
       );
 
   void _typeLink(BuildContext context) {
@@ -98,7 +100,7 @@ class ImageButton extends StatelessWidget {
       final index = controller.selection.baseOffset;
       final length = controller.selection.extentOffset - index;
 
-      controller.replaceText(index, length, BlockEmbed.image(value), null);
+      controller.replaceText(index, length, BlockEmbed.video(value), null);
     }
   }
 }
