+++ lib/src/embeds/widgets/youtube_video_app.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:youtube_player_flutter_quill/youtube_player_flutter_quill.dart';
 
-import '../default_styles.dart';
+import '../../widgets/default_styles.dart';
 
 class YoutubeVideoApp extends StatefulWidget {
   const YoutubeVideoApp(
